Fixed incorrect types on Vertex attribute

+++ source/googlemapwidget.cpp
@@ -54,7 +54,7 @@ void GoogleMapWidget::initializeGL()
 	glAttachShader(shaderProgram, fragShaderId);
 	glAttachShader(shaderProgram, vertShaderId);
 	glBindAttribLocation(shaderProgram, 0, "position");
-    glBindAttribLocation(shaderProgram, 1, "texture");
+	glBindAttribLocation(shaderProgram, 1, "texture");
 	glLinkProgram(shaderProgram);
 
 	glUseProgram(shaderProgram);
@@ -76,10 +76,10 @@ void GoogleMapWidget::initializeGL()
 	glBufferData(GL_ARRAY_BUFFER, sizeof(quadVertices), quadVertices, GL_STATIC_DRAW);
 
 	glEnableVertexAttribArray(0);
-	glVertexAttribPointer(0, 2, GL_FLOAT, GL_FALSE, 4 * sizeof(GLuint), 0);
+	glVertexAttribPointer(0, 2, GL_FLOAT, GL_FALSE, 4 * sizeof(GLfloat), 0);
 
 	glEnableVertexAttribArray(1);
-	glVertexAttribPointer(1, 2, GL_FLOAT, GL_FALSE, 4 * sizeof(GLuint), (void*)(2 * sizeof(GLuint)));
+	glVertexAttribPointer(1, 2, GL_FLOAT, GL_FALSE, 4 * sizeof(GLfloat), (void*)(2 * sizeof(GLfloat)));
 
 	textureId = bindTexture(tile);
